re-factored fixture and test files, moved sharing to PropertyDescriptorUsage -- tests don't yet compile
