+++ crates/coordinator/descriptors/tests/property_descriptor_tests.rs
@@ -17,10 +17,10 @@ use descriptors::helpers::{
 use descriptors::property_map_builder::upsert_property_descriptor;
 
 use rstest::*;
-use shared_test::fixture_defs::{
+use shared_test::property_descriptor_fixtures::{
     new_dedicated_property_descriptors_fixture, new_shared_property_descriptors_fixture,
 };
-use shared_test::SharedTypesTestCase;
+use shared_test::test_data_types::SharedTypesTestCase;
 use shared_types_descriptor::error::DescriptorsError;
 use shared_types_descriptor::property_descriptor::{
     CompositeDescriptor, DescriptorSharing, PropertyDescriptor, PropertyDescriptorDetails,
@@ -75,7 +75,7 @@ async fn rstest_property_descriptor_capabilities(
 
     for descriptor in descriptors.clone() {
         println!("Starting create/get test for the following PropertyDescriptor");
-        // println!("{:#?}", descriptor);
+        println!("{:#?}", descriptor);
 
         let created_record: Record = conductor
             .call(
@@ -132,310 +132,3 @@ async fn rstest_property_descriptor_capabilities(
     assert_eq!(descriptors, fetched_entries);
 }
 
-/// Testing shared variants of properties within a composite
-#[rstest]
-#[case::mixture_of_property_types(new_shared_property_descriptors_fixture())]
-#[tokio::test(flavor = "multi_thread")]
-async fn rstest_shared_properties(#[case] input: Result<SharedTypesTestCase, DescriptorsError>) {
-    let (conductor, _agent, cell): (SweetConductor, AgentPubKey, SweetCell) =
-        shared_test::setup_conductor().await;
-
-    println!(
-        "******* STARTING TESTS FOR SHARED PROPERTY DESCRIPTORS *************************** \n"
-    );
-
-    let test_case = input.unwrap();
-    let shared_types = test_case.shared_types;
-    let referencing_types = test_case.referencing_types;
-
-    let shared_string_name = "Shared_String_Type_example".to_string();
-    let shared_integer_name = "Shared_I64_Integer_Type_example".to_string();
-    let shared_boolean_name = "Shared_Boolean_Type_example".to_string();
-
-    let mut created_shared_types: Vec<PropertyDescriptor> = Vec::new();
-    // Create each shared type as an entry in Holochain, then collect them
-    for descriptor in shared_types {
-        let created_record: Record = conductor
-            .call(
-                &cell.zome("descriptors"),
-                "create_property_descriptor",
-                descriptor.clone(),
-            )
-            .await;
-        let mut created_descriptor = descriptor.clone();
-        let name = get_holon_reference_from_sharing(&descriptor.sharing).name;
-        created_descriptor.sharing = DescriptorSharing::Shared(HolonReference {
-            id: Some(created_record.action_address().clone()),
-            name: name,
-        });
-        created_shared_types.push(created_descriptor);
-        let fetched_record: Option<Record> = conductor
-            .call(
-                &cell.zome("descriptors"),
-                "get_property_descriptor",
-                created_record.action_address().clone(),
-            )
-            .await;
-        let fetched_descriptor =
-            get_property_descriptor_from_record(fetched_record.unwrap()).unwrap();
-        assert_eq!(descriptor, fetched_descriptor);
-    }
-
-    let created_shared_string_type: PropertyDescriptor = created_shared_types
-        .iter()
-        .find(|descriptor| descriptor.header.type_name == shared_string_name)
-        .unwrap()
-        .clone();
-
-    let created_shared_integer_type: PropertyDescriptor = created_shared_types
-        .iter()
-        .find(|descriptor| descriptor.header.type_name == shared_integer_name)
-        .unwrap()
-        .clone();
-
-    let created_shared_boolean_type: PropertyDescriptor = created_shared_types
-        .iter()
-        .find(|descriptor| descriptor.header.type_name == shared_boolean_name)
-        .unwrap()
-        .clone();
-
-    // println!(
-    //     "string: {:#?}, \n int: {:#?}, \n bool: {:#?}",
-    //     shared_string_tuple, shared_integer_tuple, shared_boolean_tuple
-    // );
-    // println!("referencing types: {:#?}", referencing_types);
-
-    // TESTING SHARED STRING TYPE //
-
-    // Iterate referencing types to find TestComposite1
-    let mut composite1_referencing_string_descriptor: PropertyDescriptor = referencing_types
-        .iter()
-        .find(|descriptor| {
-            descriptor.header.type_name
-                == "TestComposite1__Composite_Type__referencing_string".to_string()
-        })
-        .unwrap()
-        .clone();
-    // Get the associated map, then upsert the name and usage
-    let mut composite1_map: PropertyDescriptorMap =
-        get_composite_descriptor_map(&composite1_referencing_string_descriptor.details);
-    let composite1_usage = PropertyDescriptorUsage::new(
-        "testing reference to shared string type".to_string(),
-        created_shared_string_type.clone(),
-    );
-    upsert_property_descriptor(
-        &mut composite1_map,
-        shared_string_name.clone(),
-        &composite1_usage,
-    );
-    // Update the details
-    composite1_referencing_string_descriptor.details =
-        PropertyDescriptorDetails::Composite(CompositeDescriptor {
-            properties: composite1_map,
-        });
-    // println!("{:#?}", composite1_referencing_string_descriptor);
-
-    // Create the referencing Composite type, then fetch it and compare
-    let created_composite1_record: Record = conductor
-        .call(
-            &cell.zome("descriptors"),
-            "create_property_descriptor",
-            composite1_referencing_string_descriptor.clone(),
-        )
-        .await;
-    let fetched_composite1_record: Option<Record> = conductor
-        .call(
-            &cell.zome("descriptors"),
-            "get_property_descriptor",
-            created_composite1_record.action_address().clone(),
-        )
-        .await;
-    let fetched_composite1 =
-        get_property_descriptor_from_record(fetched_composite1_record.unwrap()).unwrap();
-    assert_eq!(composite1_referencing_string_descriptor, fetched_composite1);
-    // Retrieve the ActionHash of the shared string type from the HolonReference, then fetch the property and compare
-    let referenced_string_sharing =
-        get_composite_descriptor_from_details(&fetched_composite1.details)
-            .properties
-            .properties
-            .get(&shared_string_name)
-            .unwrap()
-            .descriptor
-            .sharing
-            .clone();
-    let referenced_string_action_hash =
-        get_holon_reference_from_sharing(&referenced_string_sharing)
-            .id
-            .unwrap();
-    let fetched_referenced_string_record: Option<Record> = conductor
-        .call(
-            &cell.zome("descriptors"),
-            "get_property_descriptor",
-            referenced_string_action_hash,
-        )
-        .await;
-    let fetched_referenced_string_descriptor =
-        get_property_descriptor_from_record(fetched_referenced_string_record.unwrap()).unwrap();
-    assert_eq!(
-        created_shared_string_type,
-        fetched_referenced_string_descriptor
-    );
-
-    // TESTING SHARED INTEGER TYPE I64 //
-
-    // Iterate referencing types to find TestComposite2
-    let mut composite2_referencing_integer_descriptor: PropertyDescriptor = referencing_types
-        .iter()
-        .find(|descriptor| {
-            descriptor.header.type_name
-                == "TestComposite2__Composite_Type__referencing_integer".to_string()
-        })
-        .unwrap()
-        .clone();
-    // Get the associated map, then upsert the name and usage
-    let mut composite2_map: PropertyDescriptorMap =
-        get_composite_descriptor_map(&composite2_referencing_integer_descriptor.details);
-    let composite2_usage = PropertyDescriptorUsage::new(
-        "testing reference to shared integer type".to_string(),
-        created_shared_integer_type.clone(),
-    );
-    upsert_property_descriptor(
-        &mut composite2_map,
-        shared_integer_name.clone(),
-        &composite2_usage,
-    );
-    // Update the details
-    composite2_referencing_integer_descriptor.details =
-        PropertyDescriptorDetails::Composite(CompositeDescriptor {
-            properties: composite2_map,
-        });
-    // println!("{:#?}", composite2_referencing_integer);
-
-    // Create the referencing Composite type, then fetch it and compare
-    let created_composite2_record: Record = conductor
-        .call(
-            &cell.zome("descriptors"),
-            "create_property_descriptor",
-            composite2_referencing_integer_descriptor.clone(),
-        )
-        .await;
-    let fetched_composite2_record: Option<Record> = conductor
-        .call(
-            &cell.zome("descriptors"),
-            "get_property_descriptor",
-            created_composite2_record.action_address().clone(),
-        )
-        .await;
-    let fetched_composite2 =
-        get_property_descriptor_from_record(fetched_composite2_record.unwrap()).unwrap();
-    assert_eq!(
-        composite2_referencing_integer_descriptor,
-        fetched_composite2
-    );
-    // Retrieve the ActionHash of the shared integer type from the HolonReference, then fetch the property and compare
-    let referenced_integer_sharing =
-        get_composite_descriptor_from_details(&fetched_composite2.details)
-            .properties
-            .properties
-            .get(&shared_integer_name)
-            .unwrap()
-            .descriptor
-            .sharing
-            .clone();
-    let referenced_integer_action_hash =
-        get_holon_reference_from_sharing(&referenced_integer_sharing)
-            .id
-            .unwrap();
-    let fetched_referenced_integer_record: Option<Record> = conductor
-        .call(
-            &cell.zome("descriptors"),
-            "get_property_descriptor",
-            referenced_integer_action_hash,
-        )
-        .await;
-    let fetched_referenced_integer_descriptor =
-        get_property_descriptor_from_record(fetched_referenced_integer_record.unwrap()).unwrap();
-    assert_eq!(
-        created_shared_integer_type,
-        fetched_referenced_integer_descriptor
-    );
-
-    // TESTING SHARED BOOLEAN TYPE //
-
-    // Iterate referencing types to find TestComposite3
-    let mut composite3_referencing_boolean_descriptor: PropertyDescriptor = referencing_types
-        .iter()
-        .find(|descriptor| {
-            descriptor.header.type_name
-                == "TestComposite1__Composite_Type__referencing_boolean".to_string()
-        })
-        .unwrap()
-        .clone();
-    // Get the associated map, then upsert the name and usage
-    let mut composite3_map: PropertyDescriptorMap =
-        get_composite_descriptor_map(&composite3_referencing_boolean_descriptor.details);
-    let composite3_usage = PropertyDescriptorUsage::new(
-        "testing reference to shared boolean type".to_string(),
-        created_shared_boolean_type.clone(),
-    );
-    upsert_property_descriptor(
-        &mut composite3_map,
-        shared_boolean_name.clone(),
-        &composite3_usage,
-    );
-    // Update the details
-    composite3_referencing_boolean_descriptor.details =
-        PropertyDescriptorDetails::Composite(CompositeDescriptor {
-            properties: composite3_map,
-        });
-    // println!("{:#?}", composite3_referencing_boolean);
-
-    // Create the referencing Composite type, then fetch it and compare
-    let created_composite3_record: Record = conductor
-        .call(
-            &cell.zome("descriptors"),
-            "create_property_descriptor",
-            composite3_referencing_boolean_descriptor.clone(),
-        )
-        .await;
-    let fetched_composite3_record: Option<Record> = conductor
-        .call(
-            &cell.zome("descriptors"),
-            "get_property_descriptor",
-            created_composite3_record.action_address().clone(),
-        )
-        .await;
-    let fetched_composite3 =
-        get_property_descriptor_from_record(fetched_composite3_record.unwrap()).unwrap();
-    assert_eq!(
-        composite3_referencing_boolean_descriptor,
-        fetched_composite3
-    );
-    // Retrieve the ActionHash of the shared boolean type from the HolonReference, then fetch the property and compare
-    let referenced_boolean_sharing =
-        get_composite_descriptor_from_details(&fetched_composite3.details)
-            .properties
-            .properties
-            .get(&shared_boolean_name)
-            .unwrap()
-            .descriptor
-            .sharing
-            .clone();
-    let referenced_boolean_action_hash =
-        get_holon_reference_from_sharing(&referenced_boolean_sharing)
-            .id
-            .unwrap();
-    let fetched_referenced_boolean_record: Option<Record> = conductor
-        .call(
-            &cell.zome("descriptors"),
-            "get_property_descriptor",
-            referenced_boolean_action_hash,
-        )
-        .await;
-    let fetched_referenced_boolean_descriptor =
-        get_property_descriptor_from_record(fetched_referenced_boolean_record.unwrap()).unwrap();
-    assert_eq!(
-        created_shared_boolean_type,
-        fetched_referenced_boolean_descriptor
-    );
-}
