+++ crates/coordinator/descriptors/tests/shared_test/holon_descriptor_fixtures.rs
@@ -12,18 +12,14 @@
 // The logic for CUD tests is identical, what varies is the test data.
 // BUT... if the test data set has all different variations in it, we may only need 1 test data set
 
-use crate::shared_test::{
-    HolonDescriptorTestCase, PropertyDescriptorTestCase, SharedTypesTestCase,
-};
+use crate::shared_test::test_data_types::HolonDescriptorTestCase;
 use core::panic;
-use descriptors::helpers::{get_composite_descriptor_from_details, get_composite_descriptor_map};
-use descriptors::mutators::{
-    new_boolean_descriptor, new_composite_descriptor, new_holon_descriptor, new_integer_descriptor,
-    new_string_descriptor,
-};
-use descriptors::property_map_builder::{remove_property_descriptor, upsert_property_descriptor};
+use std::collections::btree_map::BTreeMap;
+use descriptors::helpers::*;
+use descriptors::mutators::*;
+use descriptors::property_map_builder::*;
 use rstest::*;
-use std::collections::BTreeMap;
+
 // use hdk::prelude::*;
 use crate::shared_test::property_descriptor_data_creators::{
     create_example_property_descriptors, create_example_updates_for_property_descriptors,
@@ -35,29 +31,7 @@ use shared_types_descriptor::property_descriptor::{
     PropertyDescriptorDetails, PropertyDescriptorMap, PropertyDescriptorUsage,
 };
 use shared_types_descriptor::type_header::BaseType;
-
-pub fn derive_type_name(prefix: &str, base_type: BaseType, suffix: &str) -> String {
-    let base_type_string = base_type.to_string() + "_Type";
-    let result = if prefix.is_empty() {
-        if suffix.is_empty() {
-            base_type_string
-        } else {
-            format!("{base_type_string}_{suffix}")
-        }
-    } else {
-        // prefix is NOT empty
-        if suffix.is_empty() {
-            format!("{prefix}_{base_type_string}")
-        } else {
-            format!("{prefix}_{base_type_string}_{suffix}")
-        }
-    };
-    result.to_string()
-}
-
-// fn derive_type_description(type_name: String)-> String {
-//    format!("{type_name},_description")
-//}
+use crate::shared_test::fixture_helpers::derive_type_name;
 
 /// This function creates a rich test dataset by creating a vector of HolonDescriptors of various
 /// kinds -- from simple to complex
@@ -91,6 +65,7 @@ pub fn new_holons_fixture() -> Result<Vec<HolonDescriptor>, DescriptorsError> {
     let boolean_usage = PropertyDescriptorUsage::new(
         "example boolean usage description".to_string(),
         boolean_descriptor,
+        DescriptorSharing::default(),
     );
     upsert_property_descriptor(
         &mut holon_descriptor.properties,
@@ -130,6 +105,7 @@ pub fn new_holons_fixture() -> Result<Vec<HolonDescriptor>, DescriptorsError> {
     let composite_usage = PropertyDescriptorUsage::new(
         "example composite usage description".to_string(),
         composite_descriptor,
+        DescriptorSharing::default(),
     );
     upsert_property_descriptor(
         &mut holon_descriptor.properties,
@@ -142,182 +118,6 @@ pub fn new_holons_fixture() -> Result<Vec<HolonDescriptor>, DescriptorsError> {
     Ok(test_data_set)
 }
 
-#[fixture]
-pub fn new_dedicated_property_descriptors_fixture(
-) -> Result<Vec<PropertyDescriptor>, DescriptorsError> {
-    let mut test_data_set: Vec<PropertyDescriptor> = Vec::new();
-
-    // ----------------  PROPERTY DESCRIPTOR WITH STRING PROPERTY -------------------------------
-    let string_descriptor = new_string_descriptor(
-        derive_type_name("simple", BaseType::String, "example"),
-        "Simple Example String Property Type description".to_string(),
-        false,
-        0,
-        100,
-    )?;
-    test_data_set.push(string_descriptor);
-
-    // ----------------  PROPERTY DESCRIPTOR WITH INTEGER PROPERTY -------------------------------
-    let integer_descriptor = new_integer_descriptor(
-        derive_type_name("simple_I64", BaseType::Integer, "example"),
-        "Simple Example Integer (I64) Property Type description".to_string(),
-        false,
-        IntegerFormat::I64(),
-        -3.168e9 as i64,
-        4.44e9 as i64,
-    )?;
-    test_data_set.push(integer_descriptor);
-
-    // ----------------  PROPERTY DESCRIPTOR WITH BOOLEAN PROPERTY -------------------------------
-    let boolean_descriptor = new_boolean_descriptor(
-        derive_type_name("simple", BaseType::Boolean, "example"),
-        "Simple Example Boolean Property Type description".to_string(),
-        false,
-        false,
-    )?;
-    test_data_set.push(boolean_descriptor);
-
-    // ----------------  PROPERTY DESCRIPTOR WITH COMPOSITE PROPERTY -------------------------------
-    let mut composite_properties = PropertyDescriptorMap::new(BTreeMap::new());
-    let _unused_result = create_example_property_descriptors(&mut composite_properties)?;
-    let composite_descriptor = new_composite_descriptor(
-        derive_type_name("Simple_", BaseType::Composite, "_with_scalar_properties"),
-        "Simple Composite Property Type description".to_string(),
-        false,
-        composite_properties.clone(),
-    )?;
-    let composite_usage = PropertyDescriptorUsage::new(
-        "example composite usage description".to_string(),
-        composite_descriptor.clone(),
-    );
-    upsert_property_descriptor(
-        &mut composite_properties,
-        "a_composite_property".to_string(),
-        &composite_usage,
-    );
-    test_data_set.push(composite_descriptor);
-
-    Ok(test_data_set)
-}
-
-#[fixture]
-pub fn new_shared_property_descriptors_fixture() -> Result<SharedTypesTestCase, DescriptorsError> {
-    let mut shared_types: Vec<PropertyDescriptor> = Vec::new();
-
-    let mut string_descriptor = new_string_descriptor(
-        derive_type_name("Shared", BaseType::String, "example"),
-        "Example Shared String Property Type description".to_string(),
-        true,
-        0,
-        100,
-    )?;
-    string_descriptor.sharing = DescriptorSharing::Shared(HolonReference {
-        id: None,
-        name: Some("Shared_String_Type_example".to_string()),
-    });
-    shared_types.push(string_descriptor);
-
-    let mut integer_descriptor = new_integer_descriptor(
-        derive_type_name("Shared_I64", BaseType::Integer, "example"),
-        "Example Shared Integer (I64) Property Type description".to_string(),
-        true,
-        IntegerFormat::I64(),
-        -3.168e9 as i64,
-        4.44e9 as i64,
-    )?;
-    integer_descriptor.sharing = DescriptorSharing::Shared(HolonReference {
-        id: None,
-        name: Some("Shared_I64_Integer_Type_Example".to_string()),
-    });
-    shared_types.push(integer_descriptor);
-
-    let mut boolean_descriptor = new_boolean_descriptor(
-        derive_type_name("Shared", BaseType::Boolean, "example"),
-        "Example Shared Boolean Property Type description".to_string(),
-        true,
-        false,
-    )?;
-    boolean_descriptor.sharing = DescriptorSharing::Shared(HolonReference {
-        id: None,
-        name: Some("Shared_Boolean_Type_example".to_string()),
-    });
-    shared_types.push(boolean_descriptor);
-
-    let mut referencing_types: Vec<PropertyDescriptor> = Vec::new(); // composites
-
-    let mut composite1_properties = PropertyDescriptorMap::new(BTreeMap::new());
-    let composite_descriptor1 = new_composite_descriptor(
-        derive_type_name(
-            "TestComposite1_",
-            BaseType::Composite,
-            "_referencing_string",
-        ),
-        "Test Composite Property referencing shared string type".to_string(),
-        false,
-        composite1_properties.clone(),
-    )?;
-    let composite_usage1 = PropertyDescriptorUsage::new(
-        "testing reference from composite to shared string".to_string(),
-        composite_descriptor1.clone(),
-    );
-    upsert_property_descriptor(
-        &mut composite1_properties,
-        "test_composite_property_1".to_string(),
-        &composite_usage1,
-    );
-    referencing_types.push(composite_descriptor1);
-
-    let mut composite2_properties = PropertyDescriptorMap::new(BTreeMap::new());
-    let composite_descriptor2 = new_composite_descriptor(
-        derive_type_name(
-            "TestComposite2_",
-            BaseType::Composite,
-            "_referencing_integer",
-        ),
-        "Test Composite Property referencing shared integer type".to_string(),
-        false,
-        composite2_properties.clone(),
-    )?;
-    let composite_usage2 = PropertyDescriptorUsage::new(
-        "testing reference from composite to shared integer".to_string(),
-        composite_descriptor2.clone(),
-    );
-    upsert_property_descriptor(
-        &mut composite2_properties,
-        "test_composite_property_2".to_string(),
-        &composite_usage2,
-    );
-    referencing_types.push(composite_descriptor2);
-
-    let mut composite3_properties = PropertyDescriptorMap::new(BTreeMap::new());
-    let composite_descriptor3 = new_composite_descriptor(
-        derive_type_name(
-            "TestComposite3_",
-            BaseType::Composite,
-            "_referencing_boolean",
-        ),
-        "Test Composite Property referencing shared boolean type".to_string(),
-        false,
-        composite3_properties.clone(),
-    )?;
-    let composite_usage3 = PropertyDescriptorUsage::new(
-        "testing reference from composite to shared boolean".to_string(),
-        composite_descriptor3.clone(),
-    );
-    upsert_property_descriptor(
-        &mut composite3_properties,
-        "test_composite_property_3".to_string(),
-        &composite_usage3,
-    );
-    referencing_types.push(composite_descriptor3);
-
-    let test_case = SharedTypesTestCase {
-        shared_types,
-        referencing_types,
-    };
-
-    Ok(test_case)
-}
 
 // Builds initial HolonDescriptor with no properties
 #[fixture]
@@ -405,6 +205,7 @@ pub fn add_properties_to_composite() -> Result<HolonDescriptorTestCase, Descript
     let boolean_usage = PropertyDescriptorUsage::new(
         "new boolean usage description".to_string(),
         new_boolean_descriptor,
+        DescriptorSharing::default(),
     );
 
     let new_string_descriptor = new_string_descriptor(
@@ -417,6 +218,7 @@ pub fn add_properties_to_composite() -> Result<HolonDescriptorTestCase, Descript
     let string_usage = PropertyDescriptorUsage::new(
         "new string usage description".to_string(),
         new_string_descriptor,
+        DescriptorSharing::default(),
     );
 
     let new_i8_descriptor = new_integer_descriptor(
@@ -428,7 +230,11 @@ pub fn add_properties_to_composite() -> Result<HolonDescriptorTestCase, Descript
         42,
     )?;
     let i8_usage =
-        PropertyDescriptorUsage::new("new i8 usage description".to_string(), new_i8_descriptor);
+        PropertyDescriptorUsage::new(
+            "new i8 usage description".to_string(),
+            new_i8_descriptor,
+            DescriptorSharing::default(),
+        );
 
     let new_u64_descriptor = new_integer_descriptor(
         derive_type_name("new", BaseType::Integer, "addition"),
@@ -439,7 +245,11 @@ pub fn add_properties_to_composite() -> Result<HolonDescriptorTestCase, Descript
         999999999,
     )?;
     let u64_usage =
-        PropertyDescriptorUsage::new("new u64 usage description".to_string(), new_u64_descriptor);
+        PropertyDescriptorUsage::new(
+            "new u64 usage description".to_string(),
+            new_u64_descriptor,
+            DescriptorSharing::default(),
+        );
 
     let original_composite_property_descriptor = original_descriptor
         .properties
@@ -466,7 +276,6 @@ pub fn add_properties_to_composite() -> Result<HolonDescriptorTestCase, Descript
 
         let updated_composite_descriptor = PropertyDescriptor {
             header: new_descriptor.header,
-            sharing: DescriptorSharing::default(),
             details: PropertyDescriptorDetails::Composite(CompositeDescriptor {
                 properties: composite_descriptor_map.clone(),
             }),
@@ -474,6 +283,7 @@ pub fn add_properties_to_composite() -> Result<HolonDescriptorTestCase, Descript
         let composite_usage = PropertyDescriptorUsage::new(
             "new composite usage description".to_string(),
             updated_composite_descriptor,
+            DescriptorSharing::default(),
         );
         upsert_property_descriptor(
             &mut updated_descriptor.properties,
@@ -524,7 +334,6 @@ pub fn remove_properties_from_composite(
 
         let updated_composite_descriptor = PropertyDescriptor {
             header: usage.descriptor.header.clone(),
-            sharing: DescriptorSharing::default(),
             details: PropertyDescriptorDetails::Composite(CompositeDescriptor {
                 properties: composite_descriptor_map.clone(),
             }),
@@ -532,6 +341,7 @@ pub fn remove_properties_from_composite(
         let composite_usage = PropertyDescriptorUsage::new(
             "new composite usage description".to_string(),
             updated_composite_descriptor,
+            DescriptorSharing::default(),
         );
         upsert_property_descriptor(
             &mut updated_descriptor.properties,
@@ -551,39 +361,6 @@ pub fn remove_properties_from_composite(
     }
 }
 
-#[fixture]
-pub fn update_property_descriptor_composite() -> Result<PropertyDescriptorTestCase, DescriptorsError>
-{
-    let original_descriptor = build_property_descriptor_with_composite()?;
-    let mut updates = Vec::new();
-
-    let mut composite_descriptor =
-        get_composite_descriptor_from_details(&original_descriptor.details);
-    let mut descriptor_map = get_composite_descriptor_map(&original_descriptor.details);
-
-    let update_properties = create_example_updates_for_property_descriptors(&mut descriptor_map)?;
-
-    for (name, property) in update_properties.properties.clone() {
-        upsert_property_descriptor(&mut descriptor_map, name, &property);
-    }
-
-    composite_descriptor.properties = descriptor_map;
-
-    let updated_descriptor = PropertyDescriptor {
-        header: original_descriptor.header.clone(),
-        sharing: DescriptorSharing::default(),
-        details: PropertyDescriptorDetails::Composite(composite_descriptor),
-    };
-
-    updates.push(updated_descriptor.clone());
-
-    let test_case = PropertyDescriptorTestCase {
-        original: original_descriptor,
-        updates: updates,
-    };
-    // println!("Original & expected update: {:#?}", test_case);
-    Ok(test_case)
-}
 
 // Private local fns
 
@@ -624,6 +401,7 @@ fn build_holon_descriptor_with_composite() -> Result<HolonDescriptor, Descriptor
     let composite_usage = PropertyDescriptorUsage::new(
         "example composite usage description".to_string(),
         composite_descriptor,
+        DescriptorSharing::default(),
     );
     upsert_property_descriptor(
         &mut holon_descriptor.properties,
@@ -646,6 +424,7 @@ fn build_property_descriptor_with_composite() -> Result<PropertyDescriptor, Desc
     let composite_usage = PropertyDescriptorUsage::new(
         "new composite usage description".to_string(),
         composite_descriptor.clone(),
+        DescriptorSharing::default(),
     );
     upsert_property_descriptor(
         &mut composite_properties,
